pyproject.toml
[project]
name = "llm-synthesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "samgleason", email = "samuel.gleason@entalpic.ai" }
]
requires-python = ">=3.11"
dependencies = [
    "docling>=2.31.2",
    "dspy>=2.6.24",
    "fsspec>=2025.3.0",
    "gcsfs>=2025.3.0",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "ipython>=9.2.0",
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.2",
    "mistralai>=1.7.0",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.per-file-ignores]
"**/__init__.py" = ["F401"]
